# docs/openapi.yaml
openapi: 3.0.2
info:
  title: Prime Studio API
  version: 0.1.0
paths:
...
  /db/videos:
    get:
      summary: Get Db Videos
      description: Get videos from DynamoDB
      operationId: get_db_videos_db_videos_get
      parameters:
        - required: false
          schema:
            title: All
            type: boolean
            default: false
          name: all
          in: query
      responses:
        "200":
          description: Successful Response
          content:
            application/json:
              schema:
                title: Response Get Db Videos Db Videos Get
                type: array
                items:
                  $ref: "#/components/schemas/VideoDB"
        "422":
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
  /videos:
    get:
      summary: Get Videos
      description: Get videos merged VimeoAPI & DynamoDB
      operationId: get_videos_videos_get
      parameters:
        - required: false
          schema:
            title: All
            type: boolean
            default: false
          name: all
          in: query
      responses:
        "200":
          description: Successful Response
          content:
            application/json:
              schema:
                title: Response Get Videos Videos Get
                type: array
                items:
                  $ref: "#/components/schemas/Video"
        "422":
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
  /video/{video_id}:
    get:
      summary: Get Video
      description: Get video from DynamoDB
      operationId: get_video_video__video_id__get
      parameters:
        - required: true
          schema:
            title: Video Id
          name: video_id
          in: path
      responses:
        "200":
          description: Successful Response
          content:
            application/json:
              schema: {}
        "422":
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
  /video/status/:
    get:
      summary: Get Transcode Status
      description: Get upload status to vimeo
      operationId: get_transcode_status_video_status__get
      parameters:
        - required: true
          schema:
            title: Video Id
            type: string
          name: video_id
          in: query
      responses:
        "200":
          description: Successful Response
          content:
            application/json:
              schema: {}
        "422":
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
  /video:
    put:
      summary: Put Video
      description: Put video to dynamoDB
      operationId: put_video_video_put
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReqVideoPut"
        required: true
      responses:
        "200":
          description: Successful Response
          content:
            application/json:
              schema: {}
        "422":
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
    post:
      summary: Post Video To Db
      description: Post video to DynamoDB
      operationId: post_video_to_db_video_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReqVideoPost"
        required: true
      responses:
        "200":
          description: Successful Response
          content:
            application/json:
              schema: {}
        "422":
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
  /video/thumbnail/{video_id}:
    post:
      summary: Post Thumbnail Vimeo
      description: Upload thumbnail to vimeo
      operationId: post_thumbnail_vimeo_video_thumbnail__video_id__post
      parameters:
        - required: true
          schema:
            title: Video Id
            type: integer
          name: video_id
          in: path
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: >-
                #/components/schemas/Body_post_thumbnail_vimeo_video_thumbnail__video_id__post
        required: true
      responses:
        "200":
          description: Successful Response
          content:
            application/json:
              schema: {}
        "422":
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
...
  /videos/{category_id}:
    get:
      summary: Get Videos Contain Any Category
      description: get videos contain any tag
      operationId: get_videos_contain_any_category_videos__category_id__get
      parameters:
        - required: true
          schema:
            title: Category Id
          name: category_id
          in: path
      responses:
        "200":
          description: Successful Response
          content:
            application/json:
              schema: {}
        "422":
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
...
components:
  schemas:
...
    Body_post_thumbnail_vimeo_video_thumbnail__video_id__post:
      title: Body_post_thumbnail_vimeo_video_thumbnail__video_id__post
      required:
        - image
      type: object
      properties:
        image:
          title: Image
          type: string
          format: binary
...
    HTTPValidationError:
      title: HTTPValidationError
      type: object
      properties:
        detail:
          title: Detail
          type: array
          items:
            $ref: "#/components/schemas/ValidationError"
...
    PrivacyVimeo:
      title: PrivacyVimeo
      required:
        - view
      type: object
      properties:
        view:
          title: View
          type: string
      description: Privacy from Vimeo
...
    ReqVideoPost:
      title: ReqVideoPost
      required:
        - PK
        - user
        - description
        - categoryId
      type: object
      properties:
        PK:
          title: Pk
          type: string
        user:
          title: User
          type: string
        invalid:
          title: Invalid
          type: boolean
          default: false
        note:
          title: Note
          type: string
          default: No Remarks
        description:
          title: Description
          type: string
        learningPathIds:
          title: Learningpathids
          type: array
          items:
            type: string
        tagIds:
          title: Tagids
          type: array
          items:
            type: string
        categoryId:
          title: Categoryid
          type: string
      description: Post video request body
    ReqVideoPut:
      title: ReqVideoPut
      required:
        - PK
        - user
      type: object
      properties:
        PK:
          title: Pk
          type: string
        user:
          title: User
          type: string
        invalid:
          title: Invalid
          type: boolean
        note:
          title: Note
          type: string
        description:
          title: Description
          type: string
        learningPathIds:
          title: Learningpathids
          type: array
          items:
            type: string
        tagIds:
          title: Tagids
          type: array
          items:
            type: string
        categoryId:
          title: Categoryid
          type: string
      description: Put video request body
...
    StatsVimeo:
      title: StatsVimeo
      required:
        - plays
      type: object
      properties:
        plays:
          title: Plays
          type: integer
      description: Stats from Vimeo
...
    ThumbnailVimeo:
      title: ThumbnailVimeo
      required:
        - width
        - height
        - link
      type: object
      properties:
        width:
          title: Width
          type: integer
        height:
          title: Height
          type: integer
        link:
          title: Link
          type: string
        link_with_play_button:
          title: Link With Play Button
          type: string
      description: Thumbnail from Vimeo
...
    ValidationError:
      title: ValidationError
      required:
        - loc
        - msg
        - type
      type: object
      properties:
        loc:
          title: Location
          type: array
          items:
            type: string
        msg:
          title: Message
          type: string
        type:
          title: Error Type
          type: string
    Video:
      title: Video
      type: object
      properties:
        indexKey:
          title: Indexkey
          type: string
          default: ""
        PK:
          title: Pk
          type: string
          default: ""
        SK:
          title: Sk
          type: string
          default: ""
        createdAt:
          title: Createdat
          type: string
          default: ""
        createdUser:
          title: Createduser
          type: string
          default: ""
        updatedAt:
          title: Updatedat
          type: string
          default: ""
        updatedUser:
          title: Updateduser
          type: string
          default: ""
        invalid:
          title: Invalid
          type: boolean
          default: true
        note:
          title: Note
          type: string
          default: ""
        description:
          title: Description
          type: string
          default: ""
        learningPathIds:
          title: Learningpathids
          type: array
          items:
            type: string
          default: []
        tagIds:
          title: Tagids
          type: array
          items:
            type: string
          default: []
        categoryId:
          title: Categoryid
          type: string
          default: ""
        match:
          title: Match
          type: boolean
          default: true
        uri:
          title: Uri
          type: string
          default: ""
        name:
          title: Name
          type: string
          default: ""
        duration:
          title: Duration
          type: integer
          default: ""
        stats:
          title: Stats
          allOf:
            - $ref: "#/components/schemas/StatsVimeo"
          default: ""
        privacy:
          $ref: "#/components/schemas/PrivacyVimeo"
        html:
          title: Html
          type: string
          default: ""
        thumbnail:
          title: Thumbnail
          allOf:
            - $ref: "#/components/schemas/ThumbnailVimeo"
          default: ""
      description: Video response body
    VideoDB:
      title: VideoDB
      type: object
      properties:
        indexKey:
          title: Indexkey
          type: string
          default: ""
        PK:
          title: Pk
          type: string
          default: ""
        SK:
          title: Sk
          type: string
          default: ""
        createdAt:
          title: Createdat
          type: string
          default: ""
        createdUser:
          title: Createduser
          type: string
          default: ""
        updatedAt:
          title: Updatedat
          type: string
          default: ""
        updatedUser:
          title: Updateduser
          type: string
          default: ""
        invalid:
          title: Invalid
          type: boolean
          default: true
        note:
          title: Note
          type: string
          default: ""
        description:
          title: Description
          type: string
          default: ""
        learningPathIds:
          title: Learningpathids
          type: array
          items:
            type: string
          default: []
        tagIds:
          title: Tagids
          type: array
          items:
            type: string
          default: []
        categoryId:
          title: Categoryid
          type: string
          default: ""
        match:
          title: Match
          type: boolean
          default: true
      description: Video from DynamoDB
...